[BROKE] Discovered that thread priority not updating as needed
Turns out maintaining a list of blocked threads is a bad idea; this list is unreliable when the the lock that was waited on changes holders.
This is causng main to have some stale priority in priority-donate-one.c. Forutnately the fix should be easy - maintain a list of locks instead.